refactor: flytt EncounterDetail.tsx til src/features/encounters/ (T-15)
Ren rename -- ingen logikkendring. Importstier i encounters/[id]/page.tsx
og CampaignWorkspaceShell.tsx oppdatert til @/-aliaser.
components/-mappa (tom etter T-14) fjernet automatisk.
305 tester gronne.

Closes #171
