+++ apps/web/src/features/encounters/EncounterDetail.test.ts
