+++ apps/web/src/features/encounters/EncounterDetail.tsx
